client/requirements.txt
customtkinter==5.2.2
opencv-python==4.10.0.84
numpy==1.26.4
httpx==0.27.0
python-dotenv==1.0.1
loguru==0.7.2
pyserial==3.5
paho-mqtt==2.1.0
bcrypt==4.2.0

# Биометрия (требует Visual C++ Build Tools для Windows)
# insightface==0.7.3
# onnxruntime==1.19.0
# faiss-cpu==1.8.0
# sqlcipher3==0.5.3
# cryptography==43.0.0